+++ src/atcb-generate-pro.js
@@ -25,6 +25,7 @@ function atcb_generate_ty(host, data) {
     // in this case, we also decorate the data (again)
     data = atcb_decorate_data(data);
   }
+  // inline svg icons
   const copyIcon =
     '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" stroke-width="1.5"><path stroke-linecap="round" stroke-linejoin="round" d="M8.25 7.5V6.108c0-1.135.845-2.098 1.976-2.192.373-.03.748-.057 1.123-.08M15.75 18H18a2.25 2.25 0 002.25-2.25V6.108c0-1.135-.845-2.098-1.976-2.192a48.424 48.424 0 00-1.123-.08M15.75 18.75v-1.875a3.375 3.375 0 00-3.375-3.375h-1.5a1.125 1.125 0 01-1.125-1.125v-1.5A3.375 3.375 0 006.375 7.5H5.25m11.9-3.664A2.251 2.251 0 0015 2.25h-1.5a2.251 2.251 0 00-2.15 1.586m5.8 0c.065.21.1.433.1.664v.75h-6V4.5c0-.231.035-.454.1-.664M6.75 7.5H4.875c-.621 0-1.125.504-1.125 1.125v12c0 .621.504 1.125 1.125 1.125h9.75c.621 0 1.125-.504 1.125-1.125V16.5a9 9 0 00-9-9z" /></svg>';
   const copiedIcon =
@@ -43,26 +44,6 @@ function atcb_generate_ty(host, data) {
   atcb_set_fullsize(tyHost.querySelector('.atcb-modal-host-initialized'));
   // get data
   const tyData = data.ty;
-  // define css
-  const scopedCss = document.createElement('style');
-  if (host.host.hasAttribute('cspnonce')) {
-    scopedCss.setAttribute('nonce', host.host.getAttribute('cspnonce'));
-  }
-  scopedCss.innerText = '.ty { text-align:center; } .atcb-modal-content p { margin:0; } p.ty-pt { margin-top:1.5em; }';
-  if (tyData.type === 'share') {
-    scopedCss.innerText += '.ty-buttons { display:flex;justify-content:center;flex-wrap:wrap; } .ty .btn-flex { display:flex;align-items:center; } .atcb-modal-btn svg { height:1.5em;width:auto;fill:none;stroke:currentColor;margin-right:.5em; }';
-  }
-  if (tyData.type === 'form') {
-    scopedCss.innerText += '.ty-form { margin-top:1.5em;padding-top:1.5em;border-top:1px solid var(--modal-btn-border);text-align:left; } .ty-field+.ty-field { padding-top: 1.3em; } .ty-field-checkradio { display:flex;align-items:center; } .ty-field-checkradio input { cursor:pointer; }';
-    scopedCss.innerText += '.ty-field p { font-size:.9em; } .ty-field label { display:block;opacity:.8;font-size:.9em; } .ty-field label.ty-sublabel { opacity:.7;padding-left:.3em;cursor:pointer; }';
-    scopedCss.innerText +=
-      '.ty-field input[type="text"], .ty-field input[type="number"] { width:100%;box-sizing:border-box;padding:.7em;font-size:.9em;opacity:.8;border:1px solid var(--modal-btn-border);border-radius:var(--modal-input-border-radius); } .ty-field input[type="text"]:hover, .ty-field input[type="number"]:hover { opacity:1; } .ty-field input[type="text"]:focus, .ty-field input[type="number"]:focus { outline:none;border-color:var(--keyboard-focus); }';
-    scopedCss.innerText += '.ty-field input[type="checkbox"], .ty-field input[type="radio"] { width:1.2em;height:1.2rem;accent-color:var(--keyboard-focus); } .ty-field input[type="checkbox"]:focus, .ty-field input[type="radio"]:focus { outline-color:var(--keyboard-focus); }';
-    scopedCss.innerText +=
-      '#submit-error { display:none;font-weight:bold;color:#f44336;text-align:center;padding-top:1.5em; } .ty-form.ty-error #submit-error { display:block; } label span { font-weight:bold;color:#f44336;padding-left:2px;} .ty-field input.ty-error { border:2px solid #f44336;accent-color:#f44336; } .ty-field input.ty-error+.ty-sublabel { opacity:1;color:#f44336; }';
-    scopedCss.innerText += '#ty-form-submit { min-width:150px;margin-left:auto;display:block;margin-right:auto; }';
-  }
-  tyHost.appendChild(scopedCss);
   // set default, if type is missing required information
   if ((tyData.type === 'link' || tyData.type === 'form') && (!tyData.url || tyData.url === '' || !tyData.url.startsWith('http'))) {
     tyData.type = 'text';
@@ -72,14 +53,14 @@ function atcb_generate_ty(host, data) {
     tyData.headline = atcb_translate_hook('thank_you', data); // TODO: Add to i18n strings
   }
   // prepare content with...
-  let tyContent = '<div class="ty">';
+  let tyContent = '<div class="pro"><p id="ty-success-msg">' + atcb_translate_hook('sent_successfully', data) + '</p><div id="ty-content">'; // TODO: Add to i18n strings
   // intro text
   if (tyData.text && tyData.text !== '') {
     tyContent += atcb_rewrite_html_elements(tyData.text);
   }
   // share buttons, if type = share
   if (tyData.type === 'share') {
-    tyContent += `<p class="ty-pt ty-buttons">
+    tyContent += `<p class="pro-pt pro-share-buttons">
     <a href="mailto:?subject=${encodeURIComponent(atcb_translate_hook('label.share.email.subject', data))}&body=%0A&#10142;%20${encodeURIComponent(tyData.url)}%0A%0A" target="_blank" rel="noopener" class="atcb-modal-btn atcb-modal-btn-primary atcb-modal-btn-border btn-flex">
       ${mailIcon}
       ${atcb_translate_hook('label.share.email', data)}
@@ -99,7 +80,7 @@ function atcb_generate_ty(host, data) {
       }
       return atcb_translate_hook('submit', data); // TODO: Add to i18n strings
     })();
-    tyContent += '<div class="' + data.identifier + '-ty-form">';
+    tyContent += '<div id="' + data.identifier + '-ty-form" class="pro-form">';
     // for each field, add respective html
     let n = 0;
     let prevType = '';
@@ -108,29 +89,30 @@ function atcb_generate_ty(host, data) {
       if (!field.name || field.name === '') {
         continue;
       }
-      if (field.type === 'hidden' && field.name.toLowerCase() === 'header') {
+      if (field.type === 'hidden' && field.name.toLowerCase() === 'header' && field.default && field.default !== '') {
         header = JSON.parse(field.default);
+        tyData.fields.splice(i - 1, 1);
         continue;
       }
       if ((prevType === 'radio' && field.type !== 'radio') || prevType !== 'radio') {
         n = i;
       }
       tyData.fields[i - 1].fieldId = data.identifier + '-ty-' + i;
-      const fieldValue = field.default || '';
+      const fieldValue = field.type === 'radio' ? field.placeholder || '' : field.default || '';
       const fieldLabel = field.label || '';
-      const fieldPlaceholder = field.placeholder || '';
+      const fieldPlaceholder = field.type === 'radio' ? '' : field.placeholder || '';
       let fieldHtml = '';
       if (prevType !== 'hidden' && i === n && i !== 1) {
         fieldHtml += '</div>';
       }
       if (field.type !== 'hidden' && i === n) {
-        fieldHtml += '<div class="ty-field' + (field.type === 'checkbox' ? ' ty-field-checkradio' : '') + '">';
+        fieldHtml += '<div class="pro-field' + (field.type === 'checkbox' ? ' pro-field-checkradio' : '') + '">';
       }
       if (field.type === 'label') {
         fieldHtml += '<p>' + field.label + '</p>';
       } else {
         if (field.type === 'radio') {
-          fieldHtml += '<div class="ty-field-checkradio">';
+          fieldHtml += '<div class="pro-field-checkradio">';
         }
         // add label
         if ((field.type === 'text' || field.type === 'number') && field.label && field.label !== '') {
@@ -156,7 +138,7 @@ function atcb_generate_ty(host, data) {
           '" />';
         // add label for checkboxes and radio buttons
         if ((field.type === 'checkbox' || field.type === 'radio') && field.label && field.label !== '') {
-          fieldHtml += '<label for="' + field.fieldId + '" class="ty-sublabel">' + field.label + (field.required ? '<span>*</span>' : '') + '</label>';
+          fieldHtml += '<label for="' + field.fieldId + '" class="pro-checkradio-label">' + field.label + (field.required ? '<span>*</span>' : '') + '</label>';
         }
         if (field.type === 'radio') {
           fieldHtml += '</div>';
@@ -168,8 +150,8 @@ function atcb_generate_ty(host, data) {
     if (prevType !== 'hidden') {
       tyContent += '</div>';
     }
-    tyContent += '<p id="submit-error">' + atcb_translate_hook('label.form.required.error', data) + '</p>'; // TODO: Add to i18n strings
-    tyContent += '<p class="ty-pt"><button id="ty-form-submit" class="atcb-modal-btn atcb-modal-btn-primary atcb-modal-btn-border">' + label + '</button></p>';
+    tyContent += '<p id="submit-error"></p>';
+    tyContent += '<p class="pro-pt"><button id="pro-form-submit" class="atcb-modal-btn atcb-modal-btn-primary atcb-modal-btn-border">' + label + '</button><span id="pro-form-submitting" class="pro-waiting"><span>.</span><span>.</span><span>.</span></span></p>';
     tyContent += '</div>';
   }
   // button with url param, if provided and type = link
@@ -180,9 +162,9 @@ function atcb_generate_ty(host, data) {
       }
       return atcb_translate_hook('continue', data);
     })();
-    tyContent += '<p class="ty-pt"><a href="' + tyData.url + '" target="_blank" rel="noopener" class="atcb-modal-btn atcb-modal-btn-primary atcb-modal-btn-border">' + label + '</a></p>';
+    tyContent += '<p class="pro-pt"><a href="' + tyData.url + '" target="_blank" rel="noopener" class="atcb-modal-btn atcb-modal-btn-primary atcb-modal-btn-border">' + label + '</a></p>';
   }
-  tyContent += '</div>';
+  tyContent += '</div></div>';
   // create modal
   atcb_create_modal(tyHost, data, 'checkmark', tyData.headline, tyContent);
   // set enhanced click functionality
@@ -206,44 +188,90 @@ function atcb_generate_ty(host, data) {
   // validate and submit form, if type = form
   if (tyData.type === 'form') {
     const tyForm = tyHost.getElementById(data.identifier + '-ty-form');
-    const tyFormSubmit = tyHost.getElementById('ty-form-submit');
-    tyFormSubmit.addEventListener('click', function (e) {
+    const errorMsg = tyHost.getElementById('submit-error');
+    const tyFormSubmit = tyHost.getElementById('pro-form-submit');
+    const tyFormSubmitting = tyHost.getElementById('pro-form-submitting');
+    tyFormSubmit.addEventListener('click', async function (e) {
       e.preventDefault();
+      tyFormSubmitting.style.display = 'block';
+      tyFormSubmit.style.display = 'none';
       let valid = true;
-      // validate fields
       tyData.fields.forEach(function (field) {
-        if (field.type !== 'label') {
+        if (field.type !== 'label' && field.type !== 'radio') {
           const input = tyHost.getElementById(field.fieldId);
-          if (field.type !== 'radio' && field.type !== 'checkbox') {
+          if (field.type !== 'checkbox') {
             input.value = atcb_secure_content(input.value.trim());
             if (field.type === 'number') {
               input.value = input.value.replace(/\D/g, '');
             }
             if (field.required && input.value === '') {
-              input.classList.add('ty-error');
+              input.classList.add('error');
+              errorMsg.textContent = atcb_translate_hook('label.form.error.required', data);
               valid = false;
             } else {
-              input.classList.remove('ty-error');
+              input.classList.remove('error');
             }
           } else {
             if (field.required && input.checked === false) {
-              input.classList.add('ty-error');
+              input.classList.add('error');
+              errorMsg.textContent = atcb_translate_hook('label.form.error.required', data);
               valid = false;
             } else {
-              input.classList.remove('ty-error');
+              input.classList.remove('error');
             }
+          }
+        }
+        if (field.type === 'radio') {
+          const radioGroup = tyHost.querySelectorAll('[name="' + field.name + '"]');
+          let checked = false;
+          radioGroup.forEach(function (radio) {
+            if (radio.checked) {
+              checked = true;
+            }
+          });
+          if (checked === false) {
+            radioGroup[0].classList.add('error');
+            errorMsg.textContent = atcb_translate_hook('label.form.error.required', data);
+            valid = false;
+          } else {
+            radioGroup[0].classList.remove('error');
           }
         }
       });
       // submit data
       if (valid) {
-        const request = sendPostRequest(tyHost, tyData.url, tyData.fields, header);
-        if (request) {
-          tyForm.classList.add('ty-success');
+        const data = [];
+        let skipRadio = false;
+        tyData.fields.forEach((field) => {
+          // push fields to data array except for labels - for radio buttons, we only push the checked one
+          if (field.type !== 'label') {
+            if (field.type === 'radio') {
+              if (!skipRadio) {
+                const radioGroup = tyHost.querySelectorAll('[name="' + field.name + '"]');
+                radioGroup.forEach(function (radio) {
+                  if (radio.checked) {
+                    data.push({ name: field.name, value: radio.value });
+                  }
+                });
+                skipRadio = true;
+              }
+            } else {
+              data.push({ name: field.name, value: tyHost.getElementById(field.fieldId).value });
+              skipRadio = false;
+            }
+          }
+        });
+        const request = await sendPostRequest(tyData.url, data, header);
+        if (request === true) {
+          tyHost.getElementById('ty-success-msg').style.display = 'block';
+          tyHost.getElementById('ty-content').style.display = 'none';
+          return;
         }
-      } else {
-        tyForm.classList.add('ty-error');
+        errorMsg.textContent = atcb_translate_hook('label.form.error.sending', data);
       }
+      tyForm.classList.add('form-error');
+      tyFormSubmitting.style.display = 'none';
+      tyFormSubmit.style.display = 'block';
     });
     tyFormSubmit.addEventListener('keyup', function (event) {
       if (event.key === 'Enter') {
@@ -255,38 +283,29 @@ function atcb_generate_ty(host, data) {
 }
 
 // FUNCTION TO SEND A REQUEST TO THE SERVER
-async function sendPostRequest(host, url, fields, header = {}) {
+async function sendPostRequest(url, fields, header = {}) {
   let formData = new FormData();
   let data = {};
+  let requestData = {};
   if (Object.keys(header).length === 0) {
     // if there is no header information, we use FormData
     fields.forEach((field) => {
-      console.log(field.fieldId);
-      let inputElement = host.getElementById(field.fieldId);
-      if (inputElement) {
-        formData.append(field.name, inputElement.value);
-      }
-      console.log(formData);
+      formData.append(field.name, field.value);
     });
+    requestData = { method: 'POST', body: formData };
   } else {
     // otherwise, we prepare and send data as JSON
     header['Content-Type'] = 'application/json';
     fields.forEach((field) => {
-      let inputElement = host.getElementById(field.fieldId);
-      if (inputElement) {
-        data[field.name] = inputElement.value;
-      }
+      data[field.name] = field.value;
     });
+    requestData = { method: 'POST', headers: header, body: JSON.stringify(data) };
   }
   // Send the FormData object using fetch
   try {
-    const response = await fetch(url, {
-      method: 'POST',
-      headers: header,
-      body: Object.keys(data).length === 0 ? formData : JSON.stringify(data),
-    });
+    const response = await fetch(url, requestData);
     if (!response.ok) {
-      throw new Error('Network response was not ok ' + response.statusText);
+      throw new Error('Network response was not ok');
     }
     return true;
   } catch (error) {
@@ -308,7 +327,7 @@ function atcb_generate_rsvp(host, data, keyboardTrigger = false, inline = false)
   /*
   On load:
   if button and expired, show
-  if button and booked out, show
+  if inline and booked out, show
   if inline, show expired inline
   if modal, pull stats and show booked out if booked out
   Render form.